requirements.txt
# Microsoft authentication & Graph API
msal==1.31.0
requests==2.32.3

# Scheduling
APScheduler==3.10.4

# Config
PyYAML==6.0.2

# Excel export
openpyxl==3.1.5

# Invoice classification
anthropic==0.40.0
pdfplumber==0.11.4
